+++ tests/About.test.ps1
@@ -6,11 +6,15 @@
 task scripts {
 	Set-Location ..
 	foreach($_ in Get-ChildItem *-*.ps1) {
-		$name = $_.FullName
+		$name = $_.Name
 
 		# has repo link
 		$r = Get-Help $name -Full
-		assert (@($r.relatedLinks)[0].navigationLink.uri -eq 'https://github.com/nightroman/PowerShelf') "$name : Expected '.Link https://github.com/nightroman/PowerShelf'."
+		$url = @($r.relatedLinks)[0].navigationLink.uri
+		assert -Message "$name : Unexpected .Link: $url" (
+			$url -eq 'https://github.com/nightroman/PowerShelf' -or
+			$url -eq "https://github.com/nightroman/PowerShelf/blob/main/docs/$name.md"
+		)
 
 		# has author info
 		if ($r.Synopsis -cnotlike '*Author: Roman Kuzmin*') {